% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{extract_network}
\alias{extract_network}
\title{Return stream network upstream of a waterbody}
\usage{
extract_network(lon = NA, lat = NA, maxsteps = 3,
  approve_all_dl = FALSE)
}
\arguments{
\item{lon}{numeric decimal degree longitude}

\item{lat}{numeric decimal degree latitude}

\item{maxsteps}{maximum number of stream climbing iterations}

\item{approve_all_dl}{logical blanket approval to download all missing data}
}
\description{
Return stream network upstream of a waterbody
}
\examples{
\dontrun{
library(mapview)
library(sf)

coords <- data.frame(lat = 20.79722, lon = -156.47833)
res <- extract_network(coords$lon, coords$lat, maxsteps = 9)

mapview(res)
}
}
