% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_state}
\alias{find_state}
\title{find_state}
\usage{
find_state(pnt, abb = FALSE)
}
\description{
find_state
}
\examples{
\dontrun{
pnt <- st_as_sf(data.frame(Lon = -107.2, Lat = 39.45),
  coords = c("Lon", "Lat"), crs = 4326)
}
}
