% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{great_lakes}
\alias{great_lakes}
\title{Data and spatial polygons of the Great Lakes}
\usage{
great_lakes(spatial = FALSE)
}
\arguments{
\item{spatial}{logical, return Great Lakes polygons?}
}
\description{
Data and spatial polygons of the Great Lakes
}
\examples{
gl <- great_lakes()
\dontrun{
gl <- great_lakes(spatial = TRUE)
}
}
