% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{data}
\name{gull}
\alias{gull}
\title{List of simple features lake polygons and flowlines within a buffer around Gull Lake Michigan.}
\description{
Data from NHD Plus
}
\details{
gull
}
\keyword{datasets}

