% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{data}
\name{gull_flow}
\alias{gull_flow}
\title{Flowlines within a buffer around Gull Lake Michigan including flow information.}
\description{
Data from NHD Plus
}
\details{
gull_flow
}
\keyword{datasets}
