% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{data}
\name{mendota}
\alias{mendota}
\title{List of simple features lake polygons and flowlines within a buffer
around Lake Mendota.}
\description{
Data from NHD Plus
}
\details{
mendota
}
\keyword{datasets}
