% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{data}
\name{mendota_flow}
\alias{mendota_flow}
\title{Flowlines within a buffer around Lake Mendotan including flow information.}
\description{
Data from NHD Plus
}
\details{
mendota_flow
}
\keyword{datasets}
