% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nhd_dl_state}
\alias{nhd_dl_state}
\title{nhd_dl_state}
\usage{
nhd_dl_state(
  state,
  state_exists,
  yes_dl,
  file_ext,
  dsn = NA,
  wkt_filter = NA,
  ...
)
}
\description{
nhd_dl_state
}
\examples{
\dontrun{
nhd_dl_state("RI", 1, 0, NA, "NHDWaterbody")
}
}
