% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nhd_get}
\alias{nhd_get}
\title{Download and cache NHD data by state}
\usage{
nhd_get(state = NA, force_dl = FALSE, force_unzip = FALSE)
}
\arguments{
\item{state}{character state abbrevation includes "DC", "PR", and "VI"}

\item{force_dl}{logical force a re-download of the requested data}

\item{force_unzip}{logical force an unzip of downloaded data}
}
\description{
Download and cache NHD data by state
}
\examples{
\dontrun{
nhd_get(state = c("DC"))
nhd_get(state = c("RI", "CT"))
}
}
