% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nhd_plus_get}
\alias{nhd_plus_get}
\title{Download and cache NHDplus data by state}
\usage{
nhd_plus_get(vpu = NA, component = "NHDSnapshot")
}
\arguments{
\item{vpu}{numeric vector processing unit}

\item{component}{character component name}
}
\description{
Download and cache NHDplus data by state
}
\examples{
\dontrun{
nhd_plus_get(vpu = 4)
nhd_plus_get(vpu = "10L")
nhd_plus_get(vpu = 1, component = "NHDPlusAttributes")

nhd_plus_get(vpu = "National", component = "V1_To_V2_Crosswalk")
}
}
