% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{nhd_plus_info}
\alias{nhd_plus_info}
\title{Return NHDplus layer metadata and field listing}
\usage{
nhd_plus_info(vpu, component, dsn, file_ext = NA)
}
\arguments{
\item{vpu}{numeric vector processing unit}

\item{component}{character component name}

\item{dsn}{character data source name}

\item{file_ext}{character choice of "shp" for spatial data and
"dbf" for non-spatial. optional}
}
\value{
An ogrinfo object from the sf package
}
\description{
Return NHDplus layer metadata and field listing
}
\examples{
\dontrun{
nhd_plus_info(vpu = 4, component = "NHDSnapshot", dsn = "NHDWaterbody")
nhd_plus_info(vpu = 1, component = "NHDPlusAttributes", dsn = "PlusFlow")
}
}
