% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{nhd_plus_list}
\alias{nhd_plus_list}
\title{List available locally cached NHDplus layers per state}
\usage{
nhd_plus_list(vpu, component = "NHDSnapshot", file_ext = NA, ...)
}
\arguments{
\item{vpu}{numeric vector processing unit}

\item{component}{character component name}

\item{file_ext}{character choice of "shp" for spatial data and
"dbf" for non-spatial. optional}

\item{...}{arguments passed to list.files. optional.}
}
\value{
A character vector of NHD layers for the specified vpu
}
\description{
List available locally cached NHDplus layers per state
}
\examples{
\dontrun{
nhd_plus_list(vpu = 4)
nhd_plus_list(vpu = 4, full.names = TRUE)

nhd_plus_list(vpu = 1, component = "NHDPlusAttributes")
nhd_plus_list(vpu = "National", component = "V1_To_V2_Crosswalk")
}
}
