% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{package}
\name{nhdR-package}
\alias{nhdR-package}
\alias{nhdR}
\title{R interface to the National Hydrography Dataset}
\description{
R interface to the National Hydrography Dataset
}
\author{
\email{stachel2@msu.edu}
}
