% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{data}
\name{sunapee}
\alias{sunapee}
\title{List of simple features lake polygons and flowlines within a buffer
around Lake Sunapee.}
\description{
Data from NHD Plus
}
\details{
sunapee
}
\keyword{datasets}
