% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdR-package.R
\docType{data}
\name{sunappe}
\alias{sunappe}
\title{List of simple features lake polygons and flowlines within a buffer
around Lake Sunappe.}
\description{
Data from NHD Plus
}
\details{
sunappe
}
\keyword{datasets}
