% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{terminal_reaches}
\alias{terminal_reaches}
\title{Return terminal reaches from collection intersecting lake}
\usage{
terminal_reaches(lon, lat)
}
\arguments{
\item{lon}{numeric decimal degree longitude}

\item{lat}{numeric decimal degree latitude}
}
\description{
Return terminal reaches from collection intersecting lake
}
\examples{
\dontrun{
coords <- data.frame(lat = 20.79722, lon = -156.47833)
terminal_reaches(coords$lon, coords$lat)

}
}
