% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toUTM}
\alias{toUTM}
\title{Re-project to appropriate UTM zone}
\usage{
toUTM(sf_object)
}
\arguments{
\item{sf_object}{an sf object}
}
\description{
Re-project to appropriate UTM zone
}
\examples{
\dontrun{
data(gull)
gull_ <- gull$sp$NHDWaterbody
st_crs(gull_)
gull_ <- st_transform(gull_, 4326)
st_crs(gull_)
st_crs(toUTM(gull_[1,]))
}
}
