% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nhd_dl_state}
\alias{nhd_dl_state}
\title{nhd_dl_state}
\usage{
nhd_dl_state(
  state,
  state_exists,
  yes_dl,
  file_ext,
  dsn = NA,
  wkt_filter = NA,
  temporary = FALSE,
  ...
)
}
\arguments{
\item{state}{state abbreviation}

\item{state_exists}{1 for file exists on disk}

\item{yes_dl}{1 for downloading the state gdb file}

\item{file_ext}{file extension ("gdb", etc)}

\item{dsn}{name of gdb layer}

\item{wkt_filter}{a text string of coordinates see sf::st_read}

\item{temporary}{logical set FALSE to save data to a persistent
rappdirs location}

\item{\dots}{other arguments passed to sf::st_read}
}
\description{
nhd_dl_state
}
\examples{
\dontrun{
nhd_dl_state("RI", 1, 0, NA, "NHDWaterbody")
}
}
