% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{nhd_info}
\alias{nhd_info}
\title{nhd_info}
\usage{
nhd_info(state, layer_name)
}
\arguments{
\item{state}{character}

\item{layer_name}{character}
}
\description{
nhd_info
}
\examples{
\dontrun{
nhd_info("DC", "NHDWaterbody")
}
}
