% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{nhd_list}
\alias{nhd_list}
\title{List available locally cached NHD layers per state}
\usage{
nhd_list(state)
}
\arguments{
\item{state}{character state abbreviation}
}
\value{
A character vector of NHD layers for the specified state
}
\description{
List available locally cached NHD layers per state
}
\examples{
\dontrun{
nhd_list(state = "DC")
}
}
