% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nhd_load}
\alias{nhd_load}
\title{nhd_load}
\usage{
nhd_load(layer_name)
}
\arguments{
\item{layer_name}{character name of a NHD layer}
}
\value{
spatial object
}
\description{
nhd_load
}
\examples{
\dontrun{
nhd_load("NHDWaterbody")
}
}

