% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{nhd_plus_get}
\alias{nhd_plus_get}
\title{nhd_plus_get}
\usage{
nhd_plus_get(vpu = NA, component = "NHDSnapshot")
}
\arguments{
\item{vpu}{numeric vector processing unit}

\item{component}{character component name}
}
\description{
nhd_plus_get
}
\examples{
\dontrun{
nhd_plus_get(vpu = 4)
}
}

