% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{nhd_plus_info}
\alias{nhd_plus_info}
\title{nhd_plus_info}
\usage{
nhd_plus_info(layer_name)
}
\arguments{
\item{layer_name}{character}
}
\description{
nhd_plus_info
}
\examples{
\dontrun{
nhd_info("Wall")
}
}

