% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{nhd_plus_list}
\alias{nhd_plus_list}
\title{nhd_plus_list}
\usage{
nhd_plus_list(fpath = NA)
}
\arguments{
\item{fpath}{file.path to nhd plus gdb file. optional.}
}
\description{
nhd_plus_list
}
\examples{
\dontrun{
nhd_plus_list()
}
}

