% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nhd_plus_load}
\alias{nhd_plus_load}
\title{nhd_plus_load}
\usage{
nhd_plus_load(vpu, component = "NHDSnapshot", dsn, file_ext = NA,
  approve_all_dl = FALSE)
}
\arguments{
\item{vpu}{numeric vector processing unit}

\item{component}{character component name}

\item{dsn}{data source name}

\item{file_ext}{character choice of "shp" for spatial data and
"dbf" for non-spatial. optional}

\item{approve_all_dl}{logical blanket approval to download all missing data}
}
\value{
spatial object
}
\description{
nhd_plus_load
}
\details{
This function will ask the user to approve downloading missing data unless approve_all_dl is set to TRUE.
}
\examples{
\dontrun{
# Spatial
dt <- nhd_plus_load(4, "NHDSnapshot", "NHDWaterbody")
dt <- nhd_plus_load(c(1,2), "NHDSnapshot", "NHDWaterbody")
dt <- nhd_plus_load(4, "NHDSnapshot", "NHDFlowline")

# Non-spatial
dt <- nhd_plus_load(1, "NHDPlusAttributes", "PlusFlow")
dt <- nhd_plus_load("National", "V1_To_V2_Crosswalk",
 "NHDPlusV1Network_V2Network_Crosswalk")
gridcode  <- nhd_plus_load(1, "NHDPlusCatchment", "featuregridcode")
}
}
