% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nhd_plus_load}
\alias{nhd_plus_load}
\title{nhd_plus_load}
\usage{
nhd_plus_load(layer_name, fpath = NA)
}
\arguments{
\item{layer_name}{character name of a NHD layer}

\item{fpath}{file.path to nhd plus gdb file. optional.}
}
\value{
spatial object
}
\description{
nhd_plus_load
}
\examples{
\dontrun{
dt <- nhd_plus_load("NHDWaterbody")
dt <- nhd_plus_load("NHDWaterbody", fpath = nhdR::gdb_plus_path())
}
}

