% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{nhd_plus_load}
\alias{nhd_plus_load}
\title{nhd_plus_load}
\usage{
nhd_plus_load(vpu, component)
}
\arguments{
\item{vpu}{numeric vector processing unit}

\item{component}{character component name}
}
\value{
spatial object
}
\description{
nhd_plus_load
}
\examples{
\dontrun{
dt <- nhd_plus_load(4, "NHDWaterbody")
dt <- nhd_plus_load(c(1,2), "NHDWaterbody")
dt <- nhd_plus_load(4, "NHDFlowline")
}
}

