% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nhd_plus_query}
\alias{nhd_plus_query}
\title{nhd_plus_query}
\usage{
nhd_plus_query(lon, lat, dsn)
}
\description{
nhd_plus_query
}
\examples{
\dontrun{
# Lake Lashaway
wk <- wikilake::lake_wiki("Gull Lake (Michigan)")
qry <- nhd_plus_query(wk$Lon, wk$Lat, dsn = "NHDWaterbody")

plot(qry$sp$geometry, col = "blue")
plot(qry$pnt, col = "red", pch = 19, add = TRUE)
axis(1); axis(2)
}
}

