% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nhd_plus_query}
\alias{nhd_plus_query}
\title{nhd_plus_query}
\usage{
nhd_plus_query(lon = NA, lat = NA, poly = NA, dsn, buffer_dist = 0.05)
}
\arguments{
\item{lon}{numeric longitude. optional}

\item{lat}{numeric latitude. optional}

\item{poly}{sfc polygon. optional}

\item{dsn}{character data source}

\item{buffer_dist}{numeric buffer in units of coordinate degrees}
}
\description{
nhd_plus_query
}
\examples{
\dontrun{
library(sf)
wk <- wikilake::lake_wiki("Gull Lake (Michigan)")

pnt <- st_as_sf(wk, coords = c("Lon", "Lat"), crs = 4326)
pnt <- st_transform(pnt, st_crs(vpu_shp))
nhd_plus_list(nhdR::find_vpu(pnt))

qry <- nhd_plus_query(wk$Lon, wk$Lat,
         dsn = c("NHDWaterbody", "NHDFlowLine"), buffer_dist = 0.05)

plot(qry$sp$NHDWaterbody$geometry, col = "blue")
plot(qry$sp$NHDFlowLine$geometry, col = "cyan", add = TRUE)
plot(qry$pnt, col = "red", pch = 19, add = TRUE)
axis(1); axis(2)

library(ggplot2)
ggplot(qry$sp$NHDWaterbody) + geom_sf()

bbox <- data.frame(xmin = -73.34628, ymin = 41.32372,
                  xmax = -73.13639, ymax = 41.66932 )
b0 <- st_sfc(st_polygon(list(rbind(c(bbox$xmin, bbox$ymin),
           c(bbox$xmax, bbox$ymin),
           c(bbox$xmax, bbox$ymax),
           c(bbox$xmin, bbox$ymax),
           c(bbox$xmin, bbox$ymin)))))
st_crs(b0) <- 4326
nhd_plus_query(poly = b0, dsn = "NHDFlowLine")
}
}
