% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nhd_plus_query}
\alias{nhd_plus_query}
\title{nhd_plus_query}
\usage{
nhd_plus_query(lon, lat, dsn, buffer_dist = 0.05)
}
\arguments{
\item{lon}{numeric longitude}

\item{lat}{numeric latitude}

\item{dsn}{character data source}

\item{buffer_dist}{numeric buffer in units of coordinate degrees}
}
\description{
nhd_plus_query
}
\examples{
\dontrun{

wk <- wikilake::lake_wiki("Webster Lake (New Hampshire)")
qry <- nhd_plus_query(wk$Lon, wk$Lat,
         dsn = c("NHDWaterbody", "NHDFlowLine"), buffer_dist = 0.02)

plot(qry$sp$NHDWaterbody$geometry, col = "blue")
plot(qry$sp$NHDFlowLine$geometry, col = "cyan", add = TRUE)
plot(qry$pnt, col = "red", pch = 19, add = TRUE)
axis(1); axis(2)
}
}

