% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{select_poly_overlay}
\alias{select_poly_overlay}
\title{Select intersect based on a polygonr}
\usage{
select_poly_overlay(poly, sp)
}
\arguments{
\item{poly}{sf *polygon object}

\item{sp}{list of sf data frames}
}
\description{
Select intersecting layer based on a polygon extent
}
