% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{terminal_reaches}
\alias{terminal_reaches}
\title{Return terminal reaches from collection intersecting lake}
\usage{
terminal_reaches(lon = NA, lat = NA, network = NA,
  approve_all_dl = FALSE)
}
\arguments{
\item{lon}{numeric decimal degree longitude}

\item{lat}{numeric decimal degree latitude}

\item{network}{sf lines collection}

\item{approve_all_dl}{logical blanket approval to download all missing data}
}
\description{
Return terminal reaches from collection intersecting lake
}
\examples{
\dontrun{
coords <- data.frame(lat = 20.79722, lon = -156.47833)
terminal_reaches(coords$lon, coords$lat)

coords <- data.frame(lat = 41.42217, lon = -73.24189)
terminal_reaches(coords$lon, coords$lat)

network <- nhd_plus_query(lon = coords$lon, lat = coords$lat,
                     dsn = "NHDFlowline", buffer_dist = 0.02)$sp$NHDFlowline
t_reach <- terminal_reaches(network = network)

plot(network$geometry)
plot(t_reach$geometry, col = "red", add = TRUE)
}
}
