% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{terminal_reaches}
\alias{terminal_reaches}
\title{Return terminal reaches from collection intersecting lake}
\usage{
terminal_reaches(lon = NA, lat = NA, network = NA,
  approve_all_dl = FALSE)
}
\arguments{
\item{lon}{numeric decimal degree longitude}

\item{lat}{numeric decimal degree latitude}

\item{network}{sf lines collection}

\item{approve_all_dl}{logical blanket approval to download all missing data}
}
\description{
In the case of a network query, a terminal reach is a stream flowline that
has no downstream reaches in-network. In the case of a point query, a
terminal reach is a flowline that exits the intersecting surface waterbody.
}
\examples{
\dontrun{
library(sf)
library(mapview)

coords  <- data.frame(lat = 20.79722, lon = -156.47833)
t_reach <- terminal_reaches(coords$lon, coords$lat)

coords  <- data.frame(lat = 41.42217, lon = -73.24189)
t_reach <- terminal_reaches(coords$lon, coords$lat)

mapview(st_as_sf(coords, coords = c("lon", "lat"), crs = 4326)) +
mapview(t_reach$geometry, color = "red")

network <- nhd_plus_query(lon = coords$lon, lat = coords$lat,
                     dsn = "NHDFlowline", buffer_dist = 0.02)$sp$NHDFlowline
t_reach <- terminal_reaches(network = network)

mapview(network) + mapview(t_reach, color = "red")
}
}
